Ubuntu booting add note for BoardConfig.py setting
If attempting to boot Ubuntu using the GRUB config
details it is necessary to add the BoardConfig.py
setting self.ENABLE_GRUB_CONFIG = 1. This setting
may not always be enabled by default.

Also add an important note about how traditional Linux
boot is only enabled with debug build of SBL or if
you have verified boot disabled. This is because there
is no means of confirming the vmlinuz, initrd, or
config.cfg as trusted files unless you have packaged
them as a signed container image.

Signed-off-by: James Gutbub <james.gutbub@intel.com>